refactor: Update Sidebar component to conditionally render profile link
This commit updates the Sidebar component in the skizify app to conditionally render the profile link based on the user's role. If the user's role is SKIZZER, the profile link is displayed. Otherwise, it is omitted. This change improves the user experience by showing relevant navigation options based on the user's role.

Note: This commit message follows the established convention of using the "refactor" type for code improvements.